requirements.txt
numpy==1.20.2
tqdm==4.51.0
python-mnist==0.7
matplotlib==3.3.3